algoforce-analytics/api/requirements.txt
fastapi>=0.118,<1.0
uvicorn>=0.30,<1.0
pandas>=2.2
numpy>=1.26
SQLAlchemy>=2.0
mysql-connector-python>=9.0
redis>=5.0
python-dotenv>=1.0
watchfiles>=1.0
